# Dynamic Memory Caps for LS, T3, T5 Objects in LST