This PR introduces dynamic memory caps for the LS, T3, and T5 objects, reducing LST's memory footprint and eliminating truncation due to underallocation.

### Memory Reduction (0.8 GeV, 10 event average)
The numbers below are the average sums over 10 events of the memory caps for all modules in an event. Previously, most of what was allocated was not used to avoid possible truncation (eg, only O(6k) T5's are created per event, but 500k elements were allocated in the T5 arrays using the constant caps). This PR fixes this approach by first doing quick counts using a subset of the cuts in the actual creation kernels to get a better estimate of how much space needs to be allocated per module.

First number is using only the static caps (ie prior to https://github.com/cms-sw/cmssw/pull/47232), second is current, third is this PR.

T5's - 502,000 -> 120,000 -> 57,000 (89% decrease)
T3's - 1,980,000 -> 737,000 -> 409,000 (79% decrease)
LS's - 1,931,000 -> 858,000 -> 532,000 (72% decrease)

### Truncation Reduction (0.8 GeV, 10 event sum)

First number is current, second is this PR. My last PR, https://github.com/cms-sw/cmssw/pull/47232, didn't change truncations compared to static caps.

T5's - 0 -> 0 (current limits are already set very high)
T3's - 1,349 -> 0
LS's - 16,488 -> 0

Timing increases on GPU by ~14% (3.6 ms -> 4.1 ms) for single stream and ~6% (1.7 ms -> 1.8 ms) for 8 streams. One thing to note however is that most of the timing increase on GPU is from reduced truncation (i.e. more objects being produced by LST, causing more computation downstream) rather than overhead from the new counting kernels that calculate dynamic caps for memory allocation. Timing with CPU backend increases by a similar 6-7%.

Master Timing
![Screenshot 2025-06-11 at 4 52 25 PM](https://github.com/user-attachments/assets/f2c9ff36-be60-45fc-adf5-4754cf8f092e)

This PR Timing
<img width="1356" height="184" alt="Screenshot 2025-08-06 at 6 50 59 PM" src="https://github.com/user-attachments/assets/982cda58-86a1-47d9-81c0-0194d297b814" />