@ariostas 2025-06-11
> run all-gpu

I changed the command to gpu-all