@GNiendorf 2025-06-11
/run gpu-all