@GNiendorf 2025-06-12
/run gpu-all