@GNiendorf 2025-07-02
It seems like moving frequently-used array references outside the inner loops improves timing pretty substantially.
![Screenshot 2025-07-02 at 2 07 02 PM](https://github.com/user-attachments/assets/02eb81de-422c-462a-9177-e2028c2840fc)