@GNiendorf 2025-07-02
Marking this is ready for review so it can be merged in soon since both @kk428 and @YonsiG need it for their own work. I will add caps on the allocations in a followup PR if that is okay, this PR works fine for ttbar and all the other samples that we use. I think @kk428 will add an additional cut in a followup PR to the T5 counting kernel to contain the T5 memory increase on the QCD sample she is using.