@GNiendorf 2025-07-02
/run gpu-all