@GNiendorf 2025-07-03
Timing increase is worse for CPU since it's serialized and the lower level objects (MDs, LSs) have some of the largest times already. The only MD cut that removes a significant amount of fakes is the dPhiChange cut which is the last cut and depends on the results before it, so I can't make a fast subset of cuts from it. Reduction in truncations also contributes. 

This PR Timing (CPU, lnx4555)
![Screenshot 2025-07-02 at 10 17 32 PM](https://github.com/user-attachments/assets/424d5592-f58f-4920-982a-014150596bf2)

Master Timing (CPU, lnx4555)
![Screenshot 2025-07-02 at 10 20 40 PM](https://github.com/user-attachments/assets/33e5572f-bcd9-4a98-8f38-db135a43b27a)