@GNiendorf 2025-07-03
Looking at the timing, it seems like the only problem is really the MD timing increase. 215->467 on CPU, accounts for like 83% of the total short increase. @slava77 Do you have any suggestions for some ways to make a subset of fast MD cuts?