@slava77 2025-07-03
> plot `hypot(dx,dy)` by itself

it should be `drt*deltaPhi` instead. the plain distance is not going to be good in the endcaps in PS layers.