@GNiendorf 2025-08-06
https://github.com/SegmentLinking/cmssw/commit/cb78027d29e10e60d2d44d4766743218efeea39a

Saving the commit before I get rid of the MD changes.