@GNiendorf 2025-08-06
/run gpu-all