@GNiendorf 2025-08-06
@slava77 Given that CPU timing isn't super important right now, do you care if I just keep the MD changes in?