@GNiendorf 2025-08-07
@slava77 Sorry for the CI spam, this PR looks good to go. One thing nice about forgetting the memset previously is that the CI showed you do get the correct caching allocator error when LST tries to request too much memory. On lnx4555 it instead fails a few lines later, so I think it is just a difference with the lnx4555 machine rather than an error with the caching allocator.