@slava77 2025-08-07
> @slava77 Given that CPU timing isn't super important right now, do you care if I just keep the MD changes in? And we speed it up in a future iteration.
> 
> Edit: I forgot to keep a memset when reverting the MD changes, caused junk values for the seg occupancies.

I don't quite agree with the CPU timing not being important. We are aiming to get LST into the HLT baseline fairly soon. CPU timing is going to be an extra hurdle and this would be making it worse.

I didn't quite follow what the problem was with reverting the MD changes. Was it just a matter of an incomplete revert or something else?