@slava77 2025-08-07
> @slava77 The MD changes were removed, so you can ignore that comment. The issue was I reverted a memset that was still needed for the segments occupancy limits. This caused junk values to be filled in the segments occupancy array. The CPU timing is fine now, similar increase to GPU.

OK, let's get this upstream then.