@GNiendorf 2025-08-25
/run gpu-all